Took out Simulation function to a separate file
